requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
pkg-resources==0.0.0
python-dotenv==0.10.1
requests==2.21.0
spotipy==2.4.4
twitter==1.18.0
urllib3==1.24.2
